Move normalization to autoPhase1
Normalization of spectrum that was done in __penalty is now
done in autoPhase1 (which calls __penalty). Normalization
condition changed. Previously sum of absolute values of
real parts of all spectrum points was to be 1, now sum of
absolute values of all (complex) spectrum points should be 1.
Should not have much effect, but needs testing.