Close files in TopSpin import
The __init__ does not close files it opens. This is now fixed for TopSpin
datatype only. Every time a file is opened, it is added to a list (self.files). In the end of the import all files in the list are closed.

Other datatypes still leave their files open - this will need fixing.
This bug can cause problems when user tries to import couple of hundred
experiments (raises "too many open files" error).