Fix line broadening issue
Line broadening caused undesired changes in integral intensity.
The real culprit was the nmrData.process method which did the line broadening
before leftshift. Now the leftshift is applied first so the line broadening
leaves the first point of the FID, which is about to be fourier transformed,
untouched. That in turn means that the integral intensity is no longer
affected by the size of line broadening.